+++ py/desitransfer/spacewatch.py
@@ -20,12 +20,6 @@ import os
 import re
 from argparse import ArgumentParser
 from html.parser import HTMLParser
-try:
-    utc = datetime.UTC
-except AttributeError:
-    # datetime.UTC is in Python 3.11
-    import pytz
-    utc = pytz.UTC
 import requests
 from desiutil.log import get_logger, DEBUG
 from . import __version__ as dtVersion
@@ -118,7 +112,7 @@ def download_jpg(files, destination, overwrite=False, test=False):
                 if r.status_code == 200:
                     downloaded += 1
                     timestamp = int(datetime.datetime.strptime(r.headers['Last-Modified'],
-                                                               '%a, %d %b %Y %H:%M:%S %Z').replace(tzinfo=utc).timestamp())
+                                                               '%a, %d %b %Y %H:%M:%S %Z').replace(tzinfo=datetime.UTC).timestamp())
                     with open(dst_jpg, 'wb') as j:
                         j.write(r.content)
                     os.utime(dst_jpg, (timestamp, timestamp))