+++ doc/changes.rst
@@ -5,11 +5,13 @@ Change Log
 1.0.6 (unreleased)
 ------------------
 
+* Update test infrastructure for Python 3.13 (PR `#71`_).
 * Ignore ``rrdetails-*.h5`` files at Utah (PR `#70`_).
 * Add additional functionality to Tucson and Utah mirror scripts (PR `#69`_).
 
 .. _`#69`: https://github.com/desihub/desitransfer/pull/69
 .. _`#70`: https://github.com/desihub/desitransfer/pull/70
+.. _`#71`: https://github.com/desihub/desitransfer/pull/71
 
 1.0.5 (2025-10-10)
 ------------------